Split EDSM into sub-modules
