+++ modules/edsm/index.js
@@ -1,0 +1,7 @@
+const getBody = require('./body');
+const getSystem = require('./system');
+
+module.exports = {
+	getBody,
+	getSystem,
+};
